requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
distlib==0.3.8
Django==5.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==25.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.13.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
mercadopago==2.2.3
msgpack==1.0.8
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
requests==2.32.3
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.2.3
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
websocket-client==1.8.0
zope.interface==6.3
